Done but throws an error
Not sure why it won't give me the error or path. I'll work on it a bit
tomorrow if I can.
Added more Todos to the READMEs
Small format change
'PI' is now a valid number
'L' (which stands for last, or the last output to a given user) is not a
valid number